requirements.txt
uvicorn
fastapi
google-generativeai
httpx